refactoring streamlit apps for importing
