+++ .vscode/settings.json
@@ -30,6 +30,7 @@
     "ASYMX",
     "ASYMY",
     "Aloney",
+    "Anonymising",
     "Astropy",
     "Ayala",
     "Behavior",
