+++ pymedphys/_streamlit/apps/dashboard.py
@@ -12,11 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-
-# pylint: disable = pointless-statement, pointless-string-statement
-# pylint: disable = no-value-for-parameter, expression-not-assigned
-# pylint: disable = too-many-lines, redefined-outer-name
-
 from pymedphys._imports import pymssql
 from pymedphys._imports import streamlit as st
