+++ pymedphys/_streamlit/apps/electrons.py
@@ -1,6 +1,16 @@
-# pylint: disable = pointless-statement, pointless-string-statement
-# pylint: disable = no-value-for-parameter, expression-not-assigned
-# pylint: disable = too-many-lines, redefined-outer-name
+# Copyright (C) 2015,2020 Cancer Care Associates
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 import re
 from glob import glob
@@ -12,148 +22,10 @@ from pymedphys._imports import streamlit as st
 
 import pymedphys.electronfactors as electronfactors
 
-"""
-# Electron Insert Factors
-"""
-
-patient_id = st.text_input("Patient ID")
-
-if patient_id == "":
-    st.stop()
-
-
-rccc_string_search_pattern = r"\\monacoda\FocalData\RCCC\1~Clinical\*~{}\plan\*\*tel.1".format(
-    patient_id
-)
-rccc_filepath_list = glob(rccc_string_search_pattern)
-
-nbccc_string_search_pattern = r"\\tunnel-nbcc-monaco\FOCALDATA\NBCCC\1~Clinical\*~{}\plan\*\*tel.1".format(
-    patient_id
-)
-nbccc_filepath_list = glob(nbccc_string_search_pattern)
-
-sash_string_search_pattern = r"\\tunnel-sash-monaco\Users\Public\Documents\CMS\FocalData\SASH\1~Clinical\*~{}\plan\*\*tel.1".format(
-    patient_id
-)
-sash_filepath_list = glob(sash_string_search_pattern)
-
-
-filepath_list = np.concatenate(
-    [rccc_filepath_list, nbccc_filepath_list, sash_filepath_list]
-)
-
-
-electronmodel_regex = r"RiverinaAgility - (\d+)MeV"
-applicator_regex = r"(\d+)X\d+"
-
-insert_data = dict()  # type: ignore
-
-for telfilepath in filepath_list:
-    insert_data[telfilepath] = dict()
-
-    with open(telfilepath, "r") as file:
-        telfilecontents = np.array(file.read().splitlines())
-
-    insert_data[telfilepath]["reference_index"] = []
-    for i, item in enumerate(telfilecontents):
-        if re.search(electronmodel_regex, item):
-            insert_data[telfilepath]["reference_index"] += [i]
-
-    insert_data[telfilepath]["applicators"] = [
-        re.search(applicator_regex, telfilecontents[i + 12]).group(1)  # type: ignore
-        for i in insert_data[telfilepath]["reference_index"]
-    ]
-
-    insert_data[telfilepath]["energies"] = [
-        re.search(electronmodel_regex, telfilecontents[i]).group(1)  # type: ignore
-        for i in insert_data[telfilepath]["reference_index"]
-    ]
-
-
-for telfilepath in filepath_list:
-    with open(telfilepath, "r") as file:
-        telfilecontents = np.array(file.read().splitlines())
-
-    insert_data[telfilepath]["x"] = []
-    insert_data[telfilepath]["y"] = []
-
-    for i, index in enumerate(insert_data[telfilepath]["reference_index"]):
-        insert_initial_range = telfilecontents[
-            index + 51 : :
-        ]  # coords start 51 lines after electron model name
-        insert_stop = np.where(insert_initial_range == "0")[0][
-            0
-        ]  # coords stop right before a line containing 0
-
-        insert_coords_string = insert_initial_range[:insert_stop]
-        insert_coords = np.fromstring(",".join(insert_coords_string), sep=",")
-        insert_data[telfilepath]["x"].append(insert_coords[0::2] / 10)
-        insert_data[telfilepath]["y"].append(insert_coords[1::2] / 10)
-
-
-for telfilepath in filepath_list:
-    insert_data[telfilepath]["width"] = []
-    insert_data[telfilepath]["length"] = []
-    insert_data[telfilepath]["circle_centre"] = []
-    insert_data[telfilepath]["P/A"] = []
-
-    for i in range(len(insert_data[telfilepath]["reference_index"])):
-
-        width, length, circle_centre = electronfactors.parameterise_insert(
-            insert_data[telfilepath]["x"][i], insert_data[telfilepath]["y"][i]
-        )
-
-        insert_data[telfilepath]["width"].append(width)
-        insert_data[telfilepath]["length"].append(length)
-        insert_data[telfilepath]["circle_centre"].append(circle_centre)
-
-        insert_data[telfilepath]["P/A"].append(
-            electronfactors.convert2_ratio_perim_area(width, length)
-        )
-
-
-data_filename = r"S:\Physics\RCCC Specific Files\Dosimetry\Elekta_EFacs\electron_factor_measured_data.csv"
-data = pd.read_csv(data_filename)
-
-width_data = data["Width (cm @ 100SSD)"]
-length_data = data["Length (cm @ 100SSD)"]
-factor_data = data["RCCC Inverse factor (dose open / dose cutout)"]
-
-p_on_a_data = electronfactors.convert2_ratio_perim_area(width_data, length_data)
-
-
-for telfilepath in filepath_list:
-    insert_data[telfilepath]["model_factor"] = []
-
-    for i in range(len(insert_data[telfilepath]["reference_index"])):
-        applicator = float(insert_data[telfilepath]["applicators"][i])
-        energy = float(insert_data[telfilepath]["energies"][i])
-        ssd = 100
-
-        reference = (
-            (data["Energy (MeV)"] == energy)
-            & (data["Applicator (cm)"] == applicator)
-            & (data["SSD (cm)"] == ssd)
-        )
-
-        number_of_measurements = np.sum(reference)
-
-        if number_of_measurements < 8:
-            insert_data[telfilepath]["model_factor"].append(np.nan)
-        else:
-            insert_data[telfilepath]["model_factor"].append(
-                electronfactors.spline_model_with_deformability(
-                    insert_data[telfilepath]["width"],
-                    insert_data[telfilepath]["P/A"],
-                    width_data[reference],
-                    p_on_a_data[reference],
-                    factor_data[reference],
-                )[0]
-            )
+# Old code warning, the below is Simon Biggs from 2015... be nice to him
 
 
 def visual_circle_and_ellipse(insert_x, insert_y, width, length, circle_centre):
-
     t = np.linspace(0, 2 * np.pi)
     circle = {
         "x": width / 2 * np.sin(t) + circle_centre[0],
@@ -204,7 +76,6 @@ def plot_insert(insert_x, insert_y, width, length, circle_centre):
 
 
 def plot_model(width_data, length_data, factor_data):
-
     i, j, k = electronfactors.create_transformed_mesh(
         width_data, length_data, factor_data
     )
@@ -235,71 +106,212 @@ def plot_model(width_data, length_data, factor_data):
     plt.ylabel("length (cm)")
 
 
-for telfilepath in filepath_list:
-    "======================================================================="
-    st.write("Filepath: `{}`".format(telfilepath))
+def main():
+    st.write(
+        """
+        # Electron Insert Factors
+        """
+    )
 
-    for i in range(len(insert_data[telfilepath]["reference_index"])):
-        applicator = float(insert_data[telfilepath]["applicators"][i])
-        energy = float(insert_data[telfilepath]["energies"][i])
-        ssd = 100
+    patient_id = st.text_input("Patient ID")
 
-        st.write("Applicator: `{} cm` | Energy: `{} MeV`".format(applicator, energy))
+    if patient_id == "":
+        st.stop()
 
-        width = insert_data[telfilepath]["width"][i]
-        length = insert_data[telfilepath]["length"][i]
+    rccc_string_search_pattern = r"\\monacoda\FocalData\RCCC\1~Clinical\*~{}\plan\*\*tel.1".format(
+        patient_id
+    )
+    rccc_filepath_list = glob(rccc_string_search_pattern)
 
-        plt.figure()
-        plot_insert(
-            insert_data[telfilepath]["x"][i],
-            insert_data[telfilepath]["y"][i],
-            insert_data[telfilepath]["width"][i],
-            insert_data[telfilepath]["length"][i],
-            insert_data[telfilepath]["circle_centre"][i],
-        )
+    nbccc_string_search_pattern = r"\\tunnel-nbcc-monaco\FOCALDATA\NBCCC\1~Clinical\*~{}\plan\*\*tel.1".format(
+        patient_id
+    )
+    nbccc_filepath_list = glob(nbccc_string_search_pattern)
 
-        reference = (
-            (data["Energy (MeV)"] == energy)
-            & (data["Applicator (cm)"] == applicator)
-            & (data["SSD (cm)"] == ssd)
-        )
+    sash_string_search_pattern = r"\\tunnel-sash-monaco\Users\Public\Documents\CMS\FocalData\SASH\1~Clinical\*~{}\plan\*\*tel.1".format(
+        patient_id
+    )
+    sash_filepath_list = glob(sash_string_search_pattern)
 
-        number_of_measurements = np.sum(reference)
+    filepath_list = np.concatenate(
+        [rccc_filepath_list, nbccc_filepath_list, sash_filepath_list]
+    )
 
-        plt.figure()
-        if number_of_measurements < 8:
-            plt.scatter(
-                width_data[reference],
-                length_data[reference],
-                s=100,
-                c=factor_data[reference],
-                cmap="viridis",
-                zorder=2,
+    electronmodel_regex = r"RiverinaAgility - (\d+)MeV"
+    applicator_regex = r"(\d+)X\d+"
+
+    insert_data = dict()  # type: ignore
+
+    for telfilepath in filepath_list:
+        insert_data[telfilepath] = dict()
+
+        with open(telfilepath, "r") as file:
+            telfilecontents = np.array(file.read().splitlines())
+
+        insert_data[telfilepath]["reference_index"] = []
+        for i, item in enumerate(telfilecontents):
+            if re.search(electronmodel_regex, item):
+                insert_data[telfilepath]["reference_index"] += [i]
+
+        insert_data[telfilepath]["applicators"] = [
+            re.search(applicator_regex, telfilecontents[i + 12]).group(
+                1
+            )  # type: ignore
+            for i in insert_data[telfilepath]["reference_index"]
+        ]
+
+        insert_data[telfilepath]["energies"] = [
+            re.search(electronmodel_regex, telfilecontents[i]).group(1)  # type: ignore
+            for i in insert_data[telfilepath]["reference_index"]
+        ]
+
+    for telfilepath in filepath_list:
+        with open(telfilepath, "r") as file:
+            telfilecontents = np.array(file.read().splitlines())
+
+        insert_data[telfilepath]["x"] = []
+        insert_data[telfilepath]["y"] = []
+
+        for i, index in enumerate(insert_data[telfilepath]["reference_index"]):
+            insert_initial_range = telfilecontents[
+                index + 51 : :
+            ]  # coords start 51 lines after electron model name
+            insert_stop = np.where(insert_initial_range == "0")[0][
+                0
+            ]  # coords stop right before a line containing 0
+
+            insert_coords_string = insert_initial_range[:insert_stop]
+            insert_coords = np.fromstring(",".join(insert_coords_string), sep=",")
+            insert_data[telfilepath]["x"].append(insert_coords[0::2] / 10)
+            insert_data[telfilepath]["y"].append(insert_coords[1::2] / 10)
+
+    for telfilepath in filepath_list:
+        insert_data[telfilepath]["width"] = []
+        insert_data[telfilepath]["length"] = []
+        insert_data[telfilepath]["circle_centre"] = []
+        insert_data[telfilepath]["P/A"] = []
+
+        for i in range(len(insert_data[telfilepath]["reference_index"])):
+
+            width, length, circle_centre = electronfactors.parameterise_insert(
+                insert_data[telfilepath]["x"][i], insert_data[telfilepath]["y"][i]
             )
-            plt.colorbar()
-        else:
-            plot_model(
-                width_data[reference], length_data[reference], factor_data[reference]
+
+            insert_data[telfilepath]["width"].append(width)
+            insert_data[telfilepath]["length"].append(length)
+            insert_data[telfilepath]["circle_centre"].append(circle_centre)
+
+            insert_data[telfilepath]["P/A"].append(
+                electronfactors.convert2_ratio_perim_area(width, length)
             )
 
-        reference_data_table = pd.concat(
-            [width_data[reference], length_data[reference], factor_data[reference]],
-            axis=1,
-        )
-        reference_data_table.sort_values(
-            ["RCCC Inverse factor (dose open / dose cutout)"],
-            ascending=False,
-            inplace=True,
-        )
+    data_filename = r"S:\Physics\RCCC Specific Files\Dosimetry\Elekta_EFacs\electron_factor_measured_data.csv"
+    data = pd.read_csv(data_filename)
 
-        reference_data_table
+    width_data = data["Width (cm @ 100SSD)"]
+    length_data = data["Length (cm @ 100SSD)"]
+    factor_data = data["RCCC Inverse factor (dose open / dose cutout)"]
 
-        st.pyplot()
+    p_on_a_data = electronfactors.convert2_ratio_perim_area(width_data, length_data)
 
-        factor = insert_data[telfilepath]["model_factor"][i]
+    for telfilepath in filepath_list:
+        insert_data[telfilepath]["model_factor"] = []
 
-        st.write(
-            "Width: `{0:0.2f} cm` | Length: `{1:0.2f} cm` | Factor: `{2:0.3f}`".format(
-                width, length, factor
+        for i in range(len(insert_data[telfilepath]["reference_index"])):
+            applicator = float(insert_data[telfilepath]["applicators"][i])
+            energy = float(insert_data[telfilepath]["energies"][i])
+            ssd = 100
+
+            reference = (
+                (data["Energy (MeV)"] == energy)
+                & (data["Applicator (cm)"] == applicator)
+                & (data["SSD (cm)"] == ssd)
             )
-        )
+
+            number_of_measurements = np.sum(reference)
+
+            if number_of_measurements < 8:
+                insert_data[telfilepath]["model_factor"].append(np.nan)
+            else:
+                insert_data[telfilepath]["model_factor"].append(
+                    electronfactors.spline_model_with_deformability(
+                        insert_data[telfilepath]["width"],
+                        insert_data[telfilepath]["P/A"],
+                        width_data[reference],
+                        p_on_a_data[reference],
+                        factor_data[reference],
+                    )[0]
+                )
+
+    for telfilepath in filepath_list:
+        st.write("---")
+        st.write("Filepath: `{}`".format(telfilepath))
+
+        for i in range(len(insert_data[telfilepath]["reference_index"])):
+            applicator = float(insert_data[telfilepath]["applicators"][i])
+            energy = float(insert_data[telfilepath]["energies"][i])
+            ssd = 100
+
+            st.write(
+                "Applicator: `{} cm` | Energy: `{} MeV`".format(applicator, energy)
+            )
+
+            width = insert_data[telfilepath]["width"][i]
+            length = insert_data[telfilepath]["length"][i]
+
+            plt.figure()
+            plot_insert(
+                insert_data[telfilepath]["x"][i],
+                insert_data[telfilepath]["y"][i],
+                insert_data[telfilepath]["width"][i],
+                insert_data[telfilepath]["length"][i],
+                insert_data[telfilepath]["circle_centre"][i],
+            )
+
+            reference = (
+                (data["Energy (MeV)"] == energy)
+                & (data["Applicator (cm)"] == applicator)
+                & (data["SSD (cm)"] == ssd)
+            )
+
+            number_of_measurements = np.sum(reference)
+
+            plt.figure()
+            if number_of_measurements < 8:
+                plt.scatter(
+                    width_data[reference],
+                    length_data[reference],
+                    s=100,
+                    c=factor_data[reference],
+                    cmap="viridis",
+                    zorder=2,
+                )
+                plt.colorbar()
+            else:
+                plot_model(
+                    width_data[reference],
+                    length_data[reference],
+                    factor_data[reference],
+                )
+
+            reference_data_table = pd.concat(
+                [width_data[reference], length_data[reference], factor_data[reference]],
+                axis=1,
+            )
+            reference_data_table.sort_values(
+                ["RCCC Inverse factor (dose open / dose cutout)"],
+                ascending=False,
+                inplace=True,
+            )
+
+            st.write(reference_data_table)
+
+            st.pyplot()
+
+            factor = insert_data[telfilepath]["model_factor"][i]
+
+            st.write(
+                "Width: `{0:0.2f} cm` | Length: `{1:0.2f} cm` | Factor: `{2:0.3f}`".format(
+                    width, length, factor
+                )
+            )
